AuthStream/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "fpdf>=1.7.2",
    "matplotlib>=3.10.7",
    "numpy>=2.3.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "streamlit>=1.50.0",
]
